backend/requirements.txt
# FastAPI and web framework
fastapi>=0.100.0,<0.115.0
uvicorn[standard]>=0.20.0,<0.32.0
python-multipart>=0.0.5

# Database
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0
alembic>=1.10.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
email-validator>=2.0.0

# HTTP client
httpx>=0.24.0
requests>=2.30.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# AI/ML
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Caching
redis>=4.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
httpx>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Production server
gunicorn>=21.2.0
